Homework file
Since ,i have not got any reply to the mail.
so i am submitting the files which have templates of login and register pages,
in which i have added get and post routes accordingly.
and then when one registers.save() method is applied.
when one logins with the credentials find(username:username,passwrod:password) is triggered.
If correce then logged in else invaild credentials.